requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - SQL Server
sqlalchemy==2.0.25
pyodbc==5.0.1
alembic==1.13.1

# Vector Database
chromadb==0.4.22

# Azure OpenAI
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Rate Limiting
slowapi==0.1.9

# Embeddings
sentence-transformers==2.3.1
tiktoken==0.5.2

# PDF Processing (for HR documents)
pypdf2==3.0.1
python-docx==1.1.0

# Monitoring (optional)
prometheus-client==0.19.0

# Async Support
aiofiles==23.2.1
asyncpg==0.29.0
